+++ js-node/nest-admin/src/abstract/paginated-result.ts
@@ -1,8 +1,0 @@
-export class PaginatedResult {
-  data: any[];
-  meta: {
-    total: number,
-    page: number,
-    last_page: number,
-  }
-}