+++ js-node/nest-admin/src/auth/auth.interceptor.spec.ts
@@ -1,7 +1,0 @@
-import { AuthInterceptor } from './auth.interceptor';
-
-describe('AuthInterceptor', () => {
-  it('should be defined', () => {
-    expect(new AuthInterceptor()).toBeDefined();
-  });
-});