+++ js-node/nest-admin/src/permission/has-permission.decorator.ts
@@ -1,3 +1,0 @@
-import { SetMetadata } from "@nestjs/common";
-
-export const HasPermission = (access: string) => SetMetadata('access', access)