+++ js-node/nest-admin/src/user/models/user-update.dto.ts
@@ -1,6 +1,0 @@
-export class UserUpdateDto {
-  first_name?: string;
-  last_name?: string;  
-  email?: string;
-  role_id?: number;
-}