+++ js-web/react-admin/src/models/permission.ts
@@ -1,8 +1,0 @@
-export class Permission {
-  constructor(
-    public id: number = 0,
-    public name: string = ''
-  ) {
-
-  }
-}