+++ js-web/react-admin/src/redux/actions/setUserAction.ts
@@ -1,8 +1,0 @@
-import { User } from "../../models/user";
-
-export const setUser = (user: User) => {
-  return {
-    type: 'SET_USER',
-    user
-  }
-}