+++ js-web/react-admin/src/redux/configureStore.ts
@@ -1,6 +1,0 @@
-import { createStore } from 'redux'
-import { setUserReducer } from './reducers/setUserReducer'
-
-export const configureStore = () => {
-  return createStore(setUserReducer)
-}